add fix to report add fix for base response

+++ src/app/Casts/ResponseCast.php
@@ -28,7 +28,7 @@ class ResponseCast implements CastsAttributes
 
         return new Response(
             new PsrResponse(
-                Arr::get($value, 'status'),
+                (int) Arr::get($value, 'status'),
                 Arr::get($value, 'headers', []),
                 json_encode(Arr::get($value, 'body'))
             ),
